Tidied up redirects
As we are now using named routes, I introduced the redirect()->route()
rather than fixed URIs to allow for flexibility of URL structure (not
that we really need this), plus it’s best practice.

I also noticed that when redirecting to a named route  using the
route() helper, it expects an array of ids, not objects.

I now use an associative array to make it explicit what i am passing
through (like ‘quiz’ => $quiz->id) but I have feeling that just the key
names are not important, and, indeed not even necessary!).

However in the interests of clarity I’ll leave it as it is.