Tidied up the models & changed question type
# Tidied up the models
Moved a lot of the create methods to the parent relation model, to
allow us to save() via the relationship, rather than saving on the
child model

# Question types
When changing question type to multi choice or diagnostic, there was no
way to define the answers. Now we have a little checker method that
monitors the change and takes the user to the answers.create view when
changing to a question type that needs answers to be defined.

# Abstracted the previews
We use the same previews of questions so I moved this about so that we
are reusing code. Also, added buttons and made the preview of the
questions look more like a preview.

# Abstracted the mutlichoice and diagnostic
Moved these into a partial (answers._answer) and call them in the
_preview views - much neater.

# Tidy up
Removed some files that are no longer being used